+++ platform/darwin/src/MGLStyleAttributeFunction_Private.h
