+++ platform/darwin/src/MGLStyle_Private.h
