+++ platform/darwin/src/NSArray+MGLStyleAttributeAdditions.mm
@@ -1,7 +1,7 @@
 #import "NSArray+MGLStyleAttributeAdditions.h"
 
-#import "NSArray+MGLStyleAttributeAdditions_Private.hpp"
-#import "MGLStyleAttributeValue_Private.hpp"
+#import "NSArray+MGLStyleAttributeAdditions_Private.h"
+#import "MGLStyleAttributeValue_Private.h"
 
 #include <array>
