+++ platform/darwin/src/NSArray+MGLStyleAttributeAdditions_Private.h
