+++ platform/darwin/src/NSColor+MGLStyleAttributeAdditions_Private.h
