+++ platform/darwin/src/NSNumber+MGLStyleAttributeAdditions_Private.h
