+++ platform/darwin/src/NSString+MGLStyleAttributeAdditions_Private.h
