+++ platform/darwin/src/NSValue+MGLStyleAttributeAdditions_Private.h
