+++ platform/darwin/src/UIColor+MGLStyleAttributeAdditions_Private.h
