+++ platform/ios/src/MGLMapView_Private.h
