+++ platform/ios/src/UIColor+MGLAdditions.h
