+++ platform/macos/src/MGLMapView_Private.h
