+++ platform/macos/src/NSColor+MGLAdditions.h
